requirements.txt
py-clob-client>=0.18.0
websocket-client>=1.6.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0

